Changes : - Added support for interrupted sync - Renamed logger to LOGGER
