+++ tap_iterable/sync.py
@@ -8,7 +8,41 @@ import singer.metrics as metrics
 from singer import metadata
 from singer import Transformer
 
-logger = singer.get_logger()
+LOGGER = singer.get_logger()
+
+def stream_is_selected(mdata):
+    return mdata.get((), {}).get('selected', False)
+
+def get_selected_streams(catalog):
+    selected_stream_names = []
+    for stream in catalog.streams:
+        mdata = metadata.to_map(stream.metadata)
+        if stream_is_selected(mdata):
+            selected_stream_names.append(stream.tap_stream_id)
+    return selected_stream_names
+
+def sync(client, catalog, state):
+    selected_stream_names = get_selected_streams(catalog)
+
+    for stream in catalog.streams:
+        stream_name = stream.tap_stream_id
+
+        mdata = metadata.to_map(stream.metadata)
+
+        if stream_name not in selected_stream_names:
+            LOGGER.info("%s: Skipping - not selected", stream_name)
+            continue
+
+        key_properties = metadata.get(mdata, (), 'table-key-properties')
+        singer.write_schema(stream_name, stream.schema.to_dict(), key_properties)
+        LOGGER.info("%s: Starting sync", stream_name)
+        instance = STREAMS[stream_name](client)
+        instance.stream = stream
+        counter_value = sync_stream(state, instance)
+        LOGGER.info("%s: Completed sync (%s rows)", stream_name, counter_value)
+    state = singer.set_currently_syncing(state, None)
+    singer.write_state(state)
+    LOGGER.info("Finished sync")
 
 
 def sync_stream(state, instance):
